k3s: reservedns variable

+++ kvirt/cluster/k3s/__init__.py
@@ -71,7 +71,7 @@ def scale(config, plandir, cluster, overrides):
 def create(config, plandir, cluster, overrides):
     platform = config.type
     data = {'kubetype': 'k3s', 'ctlplanes': 1, 'workers': 0, 'sdn': 'flannel', 'extra_scripts': [], 'autoscale': False,
-            'network': 'default', 'cloud_lb': None, 'cloud_api_internal': False}
+            'network': 'default', 'cloud_lb': None, 'cloud_api_internal': False, 'reservedns': True}
     data.update(overrides)
     data['cloud_lb'] = overrides.get('cloud_lb', platform in cloudplatforms and data['ctlplanes'] > 1)
     cloud_lb = data['cloud_lb']
@@ -82,6 +82,7 @@ def create(config, plandir, cluster, overrides):
     ctlplanes = data['ctlplanes']
     workers = data['workers']
     network = data['network']
+    reservedns = data['reservedns']
     sdn = None if 'sdn' in overrides and overrides['sdn'] is None else data.get('sdn')
     domain = data.get('domain', 'karmalabs.corp')
     image = data.get('image', 'ubuntu2004')
@@ -177,7 +178,8 @@ def create(config, plandir, cluster, overrides):
             threaded = data.get('threaded', False) or data.get('workers_threaded', False)
             config.plan(plan, inputfile=f'{plandir}/workers.yml', overrides=nodes_overrides, threaded=threaded)
     if cloud_lb and config.type in cloudplatforms:
-        config.k.delete_dns(f'api.{cluster}', domain=domain)
+        if reservedns:
+            config.k.delete_dns(f'api.{cluster}', domain=domain)
         if config.type == 'aws':
             data['vpcid'] = config.k.get_vpcid_of_vm(f"{cluster}-ctlplane-0")
         result = config.plan(plan, inputfile=f'{plandir}/cloud_lb_api.yml', overrides=data)
@@ -186,7 +188,7 @@ def create(config, plandir, cluster, overrides):
     success(f"K3s cluster {cluster} deployed!!!")
     info2(f"export KUBECONFIG=$HOME/.kcli/clusters/{cluster}/auth/kubeconfig")
     info2("export PATH=$PWD:$PATH")
-    if config.type in cloudplatforms and cloud_lb and api_ip is None:
+    if config.type in cloudplatforms and cloud_lb and reservedns:
         wait_cloud_dns(cluster, domain)
     os.environ['KUBECONFIG'] = f"{clusterdir}/auth/kubeconfig"
     apps = data.get('apps', [])
